Add tooltip notes to series+video block editor about what can be found
Should avoid some confusion in the future.
